FixFuncionDeMierdaTimestampUint64_t

+++ shared/shared/utils.c
@@ -79,7 +79,7 @@ int isNumeric (const char * s)
     return *p == '\0';
 }
 
-double getCurrentTime() {
+uint64_t getCurrentTime() {
 	struct timeval tv;
 	gettimeofday(&tv, NULL);
 	return (double) (tv.tv_sec) * 1000 + (tv.tv_usec) / 1000;
