+++ shared/shared/utils.h
@@ -6,6 +6,7 @@
 	#include <string.h>
 	#include <ctype.h>
 	#include <sys/time.h>
+	#include <inttypes.h>
 
 	// Libera memoria de un array de chars recorriendo 1 por 1
 	void split_liberar(char **);
@@ -49,6 +50,6 @@
 	/* Si el string es un número devuelve 1, sino 0.*/
 	int isNumeric (const char*);
 
-	double getCurrentTime();
+	uint64_t getCurrentTime();
 
 #endif /* SHARED_SOCKET_H_ */
